+++ src/core/composeMiddlewares.ts
@@ -1,6 +1,6 @@
-import { Context, Middleware, MiddlewareResult, PatchContext } from './interfaces';
-import { immutable } from './utils/index';
-import { combine } from './utils';
+import { Context, Middleware, MiddlewareResult, PatchContext } from '../interfaces';
+import { immutable } from '../utils/index';
+import { combine } from '../utils';
 
 /**
  * Compose middlewares to a higher-order function.
