+++ src/core/createContext.ts
@@ -1,6 +1,6 @@
-import { Querystring } from './utils/index';
-import { Context, RequestMethod, RequestOptions } from './interfaces';
-import { parseURL, mergeHeaders } from './utils';
+import { Querystring } from '../utils/index';
+import { Context, RequestMethod, RequestOptions } from '../interfaces';
+import { parseURL, mergeHeaders } from '../utils';
 
 /**
  * Create initial context by request options
