+++ src/fatcher.ts
@@ -1,8 +1,8 @@
 import { RequestOptions, Result } from './interfaces';
-import { registerMiddlewares } from './registerMiddlewares';
-import { composeMiddlewares } from './composeMiddlewares';
+import { registerMiddlewares } from './core/registerMiddlewares';
+import { composeMiddlewares } from './core/composeMiddlewares';
 import { request } from './request';
-import { createContext } from './createContext';
+import { createContext } from './core/createContext';
 import { canActivate } from './helpers';
 import { isFunction, merge } from './utils/index';
