+++ src/index.ts
@@ -1,6 +1,5 @@
 export * from './helpers';
-export { FatcherError } from './FatcherError';
+export { FatcherError } from './core/FatcherError';
 export { combine } from './utils';
 export * from './interfaces';
 export * from './fatcher';
-export * from './createScopedRequest';
